UPSTREAM: crypto: arm/aes-ce - fix for big endian
The AES key schedule generation is mostly endian agnostic, with the
exception of the rotation and the incorporation of the round constant
at the start of each round. So implement a big endian specific version
of that part to make the whole routine big endian compatible.

Fixes: 86464859cc77 ("crypto: arm - AES in ECB/CBC/CTR/XTS modes using ARMv8 Crypto Extensions")
Signed-off-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
(cherry picked from commit 58010fa6f71c9577922b22e46014b95a4ec80fa0)
Change-Id: I7eb0dc2f91c6ceba66f67a81c0d87aafdfd484a1
Signed-off-by: Eric Biggers <ebiggers@google.com>